Break out includes into _includes
Includes now exist for header, footer, categories and featured panel, allowing
for more universal reuse across the system.  Default.html is now just a call
to header.html, a row wrapper around the {{ content }}, and a call to footer.html

Also put the PVOS logo into it's own column in the grid, so it'll wrap a bit
more nicely on small screens.